Split AroonOscillatorIndicator into several files.
