+++ js/masters/indicators/aroon-oscillator.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Stock/Indicators/AroonOscillatorIndicator.js';
+import '../../Stock/Indicators/AroonOscillator/AroonOscillatorIndicator.js';
