+++ js/masters/indicators/indicators-all.src.js
@@ -16,7 +16,7 @@ import '../../Stock/Indicators/SMA/SMAIndicator.js';
 import '../../Stock/Indicators/AD/ADIndicator.js';
 import '../../Stock/Indicators/AOIndicator.js';
 import '../../Stock/Indicators/Aroon/AroonIndicator.js';
-import '../../Stock/Indicators/AroonOscillatorIndicator.js';
+import '../../Stock/Indicators/AroonOscillator/AroonOscillatorIndicator.js';
 import '../../Stock/Indicators/ATR/ATRIndicator.js';
 import '../../Stock/Indicators/BB/BBIndicator.js';
 import '../../Stock/Indicators/CCI/CCIIndicator.js';
