+++ ts/Stock/Indicators/AroonOscillator/AroonOscillatorOptions.d.ts
@@ -1,0 +1,39 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+import type {
+    AroonOptions,
+    AroonParamsOptions
+} from '../Aroon/AroonOptions';
+
+/* *
+*
+*  Declarations
+*
+* */
+
+export interface AroonOscillatorOptions
+    extends AroonOptions {
+    params?: AroonOscillatorParamsOptions;
+}
+
+export interface AroonOscillatorParamsOptions
+    extends AroonParamsOptions {
+    // for inheritance
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+export default AroonOscillatorOptions;
