+++ ts/Stock/Indicators/AroonOscillator/AroonOscillatorPoint.d.ts
@@ -1,0 +1,33 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import AroonOscillatorIndicator from './AroonOscillatorIndicator';
+import type AroonPoint from '../Aroon/AroonPoint';
+
+/* *
+*
+*  Class
+*
+* */
+
+declare class AroonOscillatorPoint extends AroonPoint {
+    public series: AroonOscillatorIndicator
+}
+
+/* *
+*
+*  Default Export
+*
+* */
+
+export default AroonOscillatorPoint;
